Fix inconsistant display number format issue in Messaging application
The issue can be reproduced as follows
1. Create  a new sms.
2. Input a recipient.
3. Edit message.
4. Add a recipient.
5. Try to edit the message
After step 5, the numbers are not formatted in the same way

The issue is that PhoneNumberFormattingTextWatcher only process the
first phone number.  This patch use ", " and "; " as separator to
split the phone numbers and then format all of phone numbers one by
one.

Change-Id: I32d59259a0cf157551e7e4b84ae6e25e68a57443
Signed-off-by: Zhang Dongxing <dongxing.zhang@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 65761
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>